Refactor end-to-end tests
